pyproject.toml
[tool.poetry]
name        = "kube-watcher"
version     = "0.1.0"
description = "Client to monitor kubernetes node activity"
authors     = ["Carlos Fernandez Musoles <carlos@kalavai.net>"]


[tool.poetry.dependencies]
python                  = ">= 3.8, < 4.0" 
kubernetes              = { version = "^27.2.0" }
prometheus-api-client   = { version = "^0.5.3" }
plotext                 = { version = "^5.2.8" }
fastapi                 = { version = "0.101.1", extras = ["all"] }
anvil-uplink            = { version = "0.4.2" }

# interface
plotly                  = { version = "^5.15.0" }


black                   = { version = ">= 22.1.0", optional = true }


[tool.poetry.extras]

dev = [
    "black"
]



[tool.poetry.dev-dependencies]
click = "*"
pre-commit = "*"
black = "*"
xenon = "*"
mypy = "*"
pydocstyle = "*"


[build-system]
requires        = [
    "poetry-core>=1.0.0",
    "setuptools"  # for pip install -e
]
build-backend   = "poetry.core.masonry.api"
